requirements.txt
torch
torchvision
transformers==4.37.2
streamlit
pillow
groq
sentencepiece